Login is now a seperate HTML, Admin toggle(Books, All and MY)
the Login from a seperate page works, Admin toggle only appears to admin users and the my and all books works and all users print to the dev tools. just need them to show on screen. Create a User creates a user saves to db  and refreshes to the index html but it doesn't recognize the user as being a user yet until they log in.